--- Corona-Warn-App/schemas/de.rki.coronawarnapp.eventregistration.checkins.storage.TraceLocationDatabase/1.json
@@ -1,112 +1,0 @@
-{
-  "formatVersion": 1,
-  "database": {
-    "version": 1,
-    "identityHash": "244eed2a4665a1c60d0792de5945888b",
-    "entities": [
-      {
-        "tableName": "checkin",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `guid` TEXT NOT NULL, `version` INTEGER NOT NULL, `type` INTEGER NOT NULL, `description` TEXT NOT NULL, `address` TEXT NOT NULL, `traceLocationStart` TEXT, `traceLocationEnd` TEXT, `defaultCheckInLengthInMinutes` INTEGER, `signature` TEXT NOT NULL, `checkInStart` TEXT NOT NULL, `checkInEnd` TEXT, `targetCheckInEnd` TEXT, `createJournalEntry` INTEGER NOT NULL)",
-        "fields": [
-          {
-            "fieldPath": "id",
-            "columnName": "id",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "guid",
-            "columnName": "guid",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "version",
-            "columnName": "version",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "type",
-            "columnName": "type",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "description",
-            "columnName": "description",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "address",
-            "columnName": "address",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "traceLocationStart",
-            "columnName": "traceLocationStart",
-            "affinity": "TEXT",
-            "notNull": false
-          },
-          {
-            "fieldPath": "traceLocationEnd",
-            "columnName": "traceLocationEnd",
-            "affinity": "TEXT",
-            "notNull": false
-          },
-          {
-            "fieldPath": "defaultCheckInLengthInMinutes",
-            "columnName": "defaultCheckInLengthInMinutes",
-            "affinity": "INTEGER",
-            "notNull": false
-          },
-          {
-            "fieldPath": "signature",
-            "columnName": "signature",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "checkInStart",
-            "columnName": "checkInStart",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "checkInEnd",
-            "columnName": "checkInEnd",
-            "affinity": "TEXT",
-            "notNull": false
-          },
-          {
-            "fieldPath": "targetCheckInEnd",
-            "columnName": "targetCheckInEnd",
-            "affinity": "TEXT",
-            "notNull": false
-          },
-          {
-            "fieldPath": "createJournalEntry",
-            "columnName": "createJournalEntry",
-            "affinity": "INTEGER",
-            "notNull": true
-          }
-        ],
-        "primaryKey": {
-          "columnNames": [
-            "id"
-          ],
-          "autoGenerate": true
-        },
-        "indices": [],
-        "foreignKeys": []
-      }
-    ],
-    "views": [],
-    "setupQueries": [
-      "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '244eed2a4665a1c60d0792de5945888b')"
-    ]
-  }
-}
